Better time formatting

+++ include/DateUtils.hpp
@@ -8,7 +8,7 @@ namespace BetterSongSearch {
         auto monthsSince = std::chrono::duration_cast<std::chrono::months>(currentEpoch - epoch2);
         return monthsSince.count();
     }
-    auto GetDateAfterMonths(int sinceEpoch, int addMonths)
+    auto GetDateAfterMonths(int64_t sinceEpoch, int64_t addMonths)
     {
         std::chrono::seconds tp(sinceEpoch);
         return tp + std::chrono::months(addMonths);
